Split up tiled_compression.py
